Refactor tests ensuring composite keys on child side into a separate test suite @ functional tests
